requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-WTF==0.14.3
Werkzeug==2.0.1
email-validator==1.1.2
Jinja2==3.0
SQLAlchemy==1.4.47
python-dotenv==1.0.0
